Add first row of footer - #3
Added the first row of the footer. This row includes the services provided by the page, divided into four columns. Each column features a descriptive icon representing the service or theme, along with a title and description. The columns are separated by vertical lines, while the row is distinguished from the rest of the footer by a horizontal line