Add responsive main - #2
Furthermore, we have implemented a maximum width feature for the main section, as well as added a container to the header. The intention behind this addition is to allow the container to expand infinitely while constraining the content within a specified maximum width. After careful consideration, we have chosen a max-width value of 1380px